backend/api/auth.py
from fastapi import APIRouter, Depends, HTTPException, Request
from sqlalchemy.orm import Session
from pydantic import BaseModel
from typing import Optional
from datetime import datetime, timedelta
from jose import jwt
import os

from core.database import get_db

router = APIRouter(prefix="/api/v1/auth", tags=["auth"])

SECRET_KEY = os.getenv("JWT_SECRET", "your-secret-key-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("JWT_EXPIRE_MINUTES", "60"))

class LoginRequest(BaseModel):
    username: str
    password: str

class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"

@router.post("/login", response_model=TokenResponse)
async def login(request: LoginRequest, db: Session = Depends(get_db)):
    # TODO: Implement user authentication
    # For now, accept any credentials (MVP placeholder)
    if not request.username or not request.password:
        raise HTTPException(status_code=400, detail="Username and password required")
    
    token_data = {
        "sub": request.username,
        "exp": datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    }
    access_token = jwt.encode(token_data, SECRET_KEY, algorithm=ALGORITHM)
    
    return TokenResponse(access_token=access_token)

@router.post("/refresh")
async def refresh_token():
    # TODO: Implement token refresh logic
    raise HTTPException(status_code=501, detail="Not implemented")

@router.post("/logout")
async def logout():
    # TODO: Implement logout logic (token blacklist)
    return {"code": 0, "message": "Logged out successfully"}
